# docs/swagger.yaml
definitions:
  api.AuthAccess:
    properties:
      identifier:
        type: string
      secret_key:
        type: string
    type: object
  api.AuthInfo:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  api.Category:
    properties:
      category:
        type: string
      sum:
        type: string
    type: object
  api.Dropdown:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  api.Event:
    properties:
      _id:
        type: string
      amount:
        type: string
      category:
        type: string
      date:
        type: string
      description:
        type: string
      type:
        type: string
    type: object
  api.HttpResponse:
    properties:
      data: {}
      is_error:
        type: boolean
      message:
        type: string
    type: object
  api.Sum:
    properties:
      categories:
        items:
          $ref: '#/definitions/api.Category'
        type: array
      sum:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
  title: Expense Tracker API
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: client information
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/api.AuthAccess'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      tags:
      - auth
  /auth/access:
    post:
      consumes:
      - application/json
      parameters:
      - description: client information
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/api.AuthAccess'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.AuthInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      tags:
      - auth
  /auth/refresh:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - auth
  /dropdown/expense:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Dropdown'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - category
  /dropdown/income:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Dropdown'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - category
  /dropdown/type:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Dropdown'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - category
  /event:
    post:
      consumes:
      - application/json
      parameters:
      - description: event information
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/api.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - event
  /event/filter:
    post:
      consumes:
      - application/json
      parameters:
      - description: filter criteria in json format
        in: body
        name: filter
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Event'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - event
  /event/month:
    get:
      consumes:
      - application/json
      parameters:
      - description: year
        in: query
        name: year
        required: true
        type: string
      - description: month to filter with
        in: query
        name: month
        required: true
        type: string
      - description: timezone
        in: query
        name: timezone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Event'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - event
  /event/sum:
    get:
      consumes:
      - application/json
      parameters:
      - description: year
        in: query
        name: year
        required: true
        type: string
      - description: month to filter with
        in: query
        name: month
        required: true
        type: string
      - description: timezone
        in: query
        name: timezone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.HttpResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.Sum'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpResponse'
      security:
      - BearerAuth: []
      tags:
      - event
  /ping:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - home
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"